fix: complete SQLAlchemy async fixture fixes and pricing contract tests
- Fixed test_user fixture to use db_session instead of get_db()
- Rebuilt PLAN_FEATURES from IndividualPlan enum for single source of truth
- Updated get_plan_features() to return complete plan details with plan_id, display_name, price_monthly_usd
- Changed VideoSession.file_size_bytes from Integer to BigInteger to support files >2GB
- Added missing coach_live_ai and coach_live_ai_advanced plans to PLAN_FEATURES

+++ backend/sql_app/models.py
@@ -1697,7 +1697,7 @@ class VideoSession(Base):
         String(500), nullable=True, comment="S3 object key for the video file"
     )
     file_size_bytes: Mapped[int | None] = mapped_column(
-        Integer, nullable=True, comment="File size in bytes for quota tracking"
+        BigInteger, nullable=True, comment="File size in bytes for quota tracking"
     )
 
     # Status tracking
